requirements.txt
flask==3.1.1
flask-cors==6.0.1
gunicorn==23.0.0
pillow==11.3.0
numpy==2.3.1
werkzeug>=3.1.0
flask
flask-cors
numpy
pillow
rembg
onnxruntime
